+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 Dev
 ===
 
+- Add ``compilestatic`` management command (replaces ``static_precompiler_watch``)
 - Add ``STATIC_PRECOMPILER_CACHE_NAME`` to settings
 
 0.6